% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.onLoad}
\alias{.onLoad}
\title{.onLoad function to set options when the package is loaded}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{a character string giving the library directory where the package defining the namespace was found.}

\item{pkgname}{a character string giving the name of the package.}
}
\description{
.onLoad function to set options when the package is loaded
}
\keyword{Internal}
