% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatastreamUserCreated_TimeSeries.R
\docType{data}
\name{DSTimeSeriesCarryIndicator}
\alias{DSTimeSeriesCarryIndicator}
\title{DSTimeSeriesCarryIndicator}
\format{
An object of class \code{list} of length 3.
}
\usage{
DSTimeSeriesCarryIndicator
}
\value{
numeric
}
\description{
This list is a supporting attribute for the CarryIndicator properties of the DSTimeSeriesRequestObject
and DSTimeSeriesResponseObjects. When you supply data which contains 'Not A Number' values (NaN) to denote non trading
days, this list instructs the mainframe in how to store the values.
}
\section{Fields}{

\describe{
\item{\code{Yes}}{Any incoming NaN values are replaced with the last non-NaN value (e.g. 1,2,3,NaN,5,NaN,7,8
will be converted and stored as 1,2,3,3,5,5,7,8).}

\item{\code{No}}{Any incoming NaN values are stored as is and returned as NaN values.}

\item{\code{Pad}}{This is similar to YES, but also pads the final value for any dates your users may request beyond the last
date in your timeseries.\cr
For example, if your timeseries supplies just 3 values 1, NaN and 3, and your user requests a range of dates two days
before and two days after your range, your user will receive the following values: \cr
No:  NaN, NaN, 1, NaN, 3, NaN, NaN \cr
Yes: NaN, NaN, 1, 1, 3, NaN, NaN \cr
Pad: NaN, NaN, 1, 1, 3, 3, 3 \cr}
}}

\keyword{datasets}
