% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatastreamUserCreated_TimeSeries.R
\docType{data}
\name{DSTimeSeriesDateAlignment}
\alias{DSTimeSeriesDateAlignment}
\title{DSTimeSeriesDateAlignment}
\format{
An object of class \code{list} of length 3.
}
\usage{
DSTimeSeriesDateAlignment
}
\value{
numeric
}
\description{
This list is a supporting attribute for the DateAlignment properties of the DSTimeSeriesRequestObject
and DSTimeSeriesResponseObjects. When you supply monthly, quarterly or annual data, the dates are stored internally
as the first day of the given period and always returned to you through this interface as the first date of the given
period. However, when your users request data from Datastream, you can specify whether the dates returned to users are
returned with dates set as the start, mid or end of the requested period
}
\section{Fields}{

\describe{
\item{\code{EndPeriod}}{This will return dates to your users that represent the last day of the month, quarter or year.}

\item{\code{StartPeriod}}{This will return dates to your users that represent the first day of the month, quarter or year.}

\item{\code{MidPeriod}}{This will return dates to your users that represent the middle of the month (15th day), quarter
(15th of the middle month) or year (30th June).}
}}

\keyword{datasets}
