% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatastreamUserCreated_TimeSeries.R
\name{DSTimeSeriesDateRangeResponse}
\alias{DSTimeSeriesDateRangeResponse}
\title{DSTimeSeriesDateRangeResponse}
\value{
DSTimeSeriesDateRangeResponse object
}
\description{
DSTimeSeriesDateRangeResponse is the object returned from the timeseries
GetTimeseriesDateRange method. This method allows you to determine the supported dates between given start
and end dates at a specified frequency.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Dates}}{A list of datetime values representing the supported dates between requested start and end
dates at a specified frequency.}

\item{\code{ResponseStatus}}{This property will contain a DSUserObjectResponseStatus value. DSUserObjectResponseStatus$UserObjectSuccess
represents a successful response.}

\item{\code{ErrorMessage}}{If ResponseStatus is not DSUserObjectResponseStatus$UserObjectSuccess this status string will provide a
description of the error condition.}

\item{\code{Properties}}{Not currently used and will currently always return NULL.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DSTimeSeriesDateRangeResponse-new}{\code{DSTimeSeriesDateRangeResponse$new()}}
\item \href{#method-DSTimeSeriesDateRangeResponse-clone}{\code{DSTimeSeriesDateRangeResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DSTimeSeriesDateRangeResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-DSTimeSeriesDateRangeResponse-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DSTimeSeriesDateRangeResponse$new(jsonDict = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{jsonDict}}{: JSON dictionary (from JSON Response)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
DSTimeSeriesDateRangeResponse object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DSTimeSeriesDateRangeResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DSTimeSeriesDateRangeResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DSTimeSeriesDateRangeResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
