% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatastreamUserCreated_TimeSeries.R
\docType{data}
\name{DSTimeSeriesFrequencyConversion}
\alias{DSTimeSeriesFrequencyConversion}
\title{DSTimeSeriesFrequencyConversion}
\format{
An object of class \code{list} of length 4.
}
\usage{
DSTimeSeriesFrequencyConversion
}
\value{
numeric
}
\description{
This list is a supporting attribute for the FrequencyConversion properties of the
DSTimeSeriesRequestObject and DSTimeSeriesResponseObjects.
This enumeration specifies how to return values if your end users requests the
timeseries at a lower frequency.
For example: if your timeseries has daily  frequency data, and your user requests monthly or
quarterly data, the FrequencyConversion property instructs the mainframe how to return monthly
or quarterly data.
}
\section{Fields}{

\describe{
\item{\code{EndValue}}{The daily value for the end of the requested period will be returned.}

\item{\code{AverageValue}}{The average of all the values for the requested period will be returned.}

\item{\code{SumValues}}{The sum of all the values for the requested period will be returned.}

\item{\code{ActualValue}}{The actual value for the requested start date will be returned for the same given date
in the requested period.}
}}

\keyword{datasets}
