% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatastreamUserCreated_TimeSeries.R
\name{DSTimeSeriesResponseObject}
\alias{DSTimeSeriesResponseObject}
\title{DSTimeSeriesResponseObject}
\value{
DSTimeSeriesResponseObject object
}
\description{
This is a subclass of DSTimeSeriesUserObjectBase and is used to return the details for a timeseries.
(See DSTimeSeriesUserObjectBase for details of all the superclass properties.)
}
\section{Super classes}{
\code{\link[DatastreamR:DSUserObjectBase]{DatastreamR::DSUserObjectBase}} -> \code{\link[DatastreamR:DSTimeSeriesUserObjectBase]{DatastreamR::DSTimeSeriesUserObjectBase}} -> \code{DSTimeSeriesResponseObject}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{DateInfo}}{A DSTimeSeriesDateInfo object defining the start date, end date and frequency
of the timeseries.}

\item{\code{DateRange}}{A DSTimeSeriesDateRange object used to return the dates and values stored in
the timeseries. See DSTimeSeriesDateRange for details.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DSTimeSeriesResponseObject-new}{\code{DSTimeSeriesResponseObject$new()}}
\item \href{#method-DSTimeSeriesResponseObject-clone}{\code{DSTimeSeriesResponseObject$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DatastreamR" data-topic="DSUserObjectBase" data-id="SetSafeUpdateParams"><a href='../../DatastreamR/html/DSUserObjectBase.html#method-DSUserObjectBase-SetSafeUpdateParams'><code>DatastreamR::DSUserObjectBase$SetSafeUpdateParams()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DSTimeSeriesResponseObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-DSTimeSeriesResponseObject-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DSTimeSeriesResponseObject$new(jsonDict = NULL, convertNullToNans = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{jsonDict}}{: JSON dictionary (from JSON Response)}

\item{\code{convertNullToNans}}{: FALSE by default, TRUE converts the NULLs in the NaNs (Not a Number)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
DSTimeSeriesResponseObject object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DSTimeSeriesResponseObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DSTimeSeriesResponseObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DSTimeSeriesResponseObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
