% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUserDataObjectBase.R
\docType{data}
\name{DSUserObjectFrequency}
\alias{DSUserObjectFrequency}
\title{DSUserObjectFrequency}
\format{
An object of class \code{list} of length 5.
}
\usage{
DSUserObjectFrequency
}
\value{
numeric
}
\description{
Regressions and Timeseries objects specify a frequency for the underlying data.
DSUserObjectFrequency defines the supported frequencies.
}
\keyword{datasets}
