% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUserDataObjectBase.R
\docType{data}
\name{DSUserObjectResponseStatus}
\alias{DSUserObjectResponseStatus}
\title{DSUserObjectResponseStatus}
\format{
An object of class \code{list} of length 6.
}
\usage{
DSUserObjectResponseStatus
}
\value{
numeric
}
\description{
All client methods to retrieve or modify user created items return a respone object which
includes a ResponseStatus property.The ResponseStatus property specifies success or failure
for the request using a DSUserObjectResponseStatus value\cr
\cr
Response Values:\cr
UserObjectSuccess : The request succeeded and the response object's UserObject(s) property
should contain the (updated) object (except for DeleteItem method).\cr
UserObjectPermissions : Users need to be specifically permissioned to create custom objects.
This flag is set if you are not currently permissioned.\cr
UserObjectNotPresent : Returned if the requested ID does not exist.\cr
UserObjectFormatError : Returned if your request object is not in the correct format.\cr
UserObjectTypeError : Returned if your supplied object is not the same as the type specified.\cr
UserObjectError : The generic error flag. This will be set for any error not specified above.\cr

Examples are: \cr
1.Requested object ID is not present \cr
2.You have exceeded the number of custom objects permitted on your account.
}
\keyword{datasets}
