% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUserDataObjectBase.R
\docType{data}
\name{DSUserObjectShareTypes}
\alias{DSUserObjectShareTypes}
\title{DSUserObjectShareTypes}
\format{
An object of class \code{list} of length 5.
}
\usage{
DSUserObjectShareTypes
}
\value{
numeric
}
\description{
All user created objects have a flag specifying how they are shared with other users.
Currently only PrivateUserGroup and Global are supported.\cr
\cr
PrivateUserGroup are items created by any Datastream ID that shares a parent Datastream ID
with your ID. Only children of the parent ID can access the user o
Expressions can be either PrivateUserGroup or Global. Like the other object types,
PrivateUserGroup items are items created by users and visible to just children of their
Datastream parent ID. PrivateUserGroup expressions have the ID signature Eaaa,
where 'a' is any alphabetical character.\cr
\cr
Global expressions are Datastream owned expressions that are available for use by any client.
These have the signature nnnE, where n is a digit.
Global expressions can be retrieved using the API service, but you cannot modify them.
Only your PrivateUserGroup items can be modified.
}
\keyword{datasets}
