% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUserDataObjectBase.R
\docType{data}
\name{DSUserObjectTypes}
\alias{DSUserObjectTypes}
\title{DSUserObjectTypes}
\format{
An object of class \code{list} of length 6.
}
\usage{
DSUserObjectTypes
}
\value{
numeric
}
\description{
Five user created types are supported. When the client classes communicate with the API
server, a DSUserObjectTypes property is set to specify the object type.
Responses from the API server also specify the type of the object being returned.
}
\keyword{datasets}
