% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetchDataRateLimited}
\alias{fetchDataRateLimited}
\title{Fetch Data with Rate Limiting}
\usage{
fetchDataRateLimited(fetchOptions)
}
\arguments{
\item{fetchOptions}{A list of fetch options including URL and query string}
}
\value{
The API response
}
\description{
\code{fetchDataRateLimited} Fetches data from the API with rate limiting.
}
\examples{
\dontrun{
library(datastreamr)
#setAPIKey("your_api_key_here")

datastreamr:::fetchDataRateLimited(
 list(
   url = "https://api.datastream.org/v1/odata/v4/Metadata",
   qs = list(`$top` = 10))
  )
  }
}
\keyword{Internal}
