% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getOptions}
\alias{getOptions}
\title{Create Request Options}
\usage{
getOptions(path, qs)
}
\arguments{
\item{path}{The endpoint path for the request}

\item{qs}{A list of query string parameters}
}
\value{
A list of request options
}
\description{
\code{getOptions} Creates the options list for making API requests.
}
\examples{
\dontrun{
library(datastreamr)
#setAPIKey("your_api_key_here")

getOptions("/v1/odata/v4/Records", list(`$select` = "Id, DatasetId"))
}
}
\keyword{Internal}
