% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastreamr.R
\name{metadata}
\alias{metadata}
\title{Get Metadata}
\usage{
metadata(qs)
}
\arguments{
\item{qs}{A list of query string parameters}
}
\value{
A dataframe of metadata results
}
\description{
\code{metadata} Retrieves metadata from the API.
}
\examples{
\dontrun{
library(datastreamr)
#setAPIKey("your_api_key_here")

metadata(list(`$select` = "Id, DatasetName"))
}
}
