% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastreamr.R
\name{partitionRequest}
\alias{partitionRequest}
\title{Handle Partitioned Requests}
\usage{
partitionRequest(path, qs)
}
\arguments{
\item{path}{The endpoint path for the request}

\item{qs}{A list of query string parameters}
}
\value{
A dataframe of partitioned results
}
\description{
\code{partitionRequest} Handles partitioned requests to the API.
}
\examples{
partitionRequest("/v1/odata/v4/Records", list(`$filter` = "LocationId eq '5717'"))
}
