% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastreamr.R
\name{records}
\alias{records}
\title{Get Records}
\usage{
records(qs)
}
\arguments{
\item{qs}{A list of query string parameters}
}
\value{
A dataframe of record results
}
\description{
\code{records} Retrieves record data from the API.
}
\examples{
\dontrun{
library(datastreamr)
#setAPIKey("your_api_key_here")

records(list(`$select` = "Id, DatasetId"))
}
}
