% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastreamr.R
\name{setAPIKey}
\alias{setAPIKey}
\title{Set API Key}
\usage{
setAPIKey(apiKeyParam)
}
\arguments{
\item{apiKeyParam}{A character string containing your unique API key}
}
\description{
\code{setAPIKey} Sets the API key for authentication.
}
\examples{
setAPIKey("your_api_key_here")
}
