% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastreamr.R
\name{fetchData}
\alias{fetchData}
\title{Fetch Data}
\usage{
fetchData(fetchOptions)
}
\arguments{
\item{fetchOptions}{A list of fetch options including URL and query string}
}
\value{
A list of results from the API
}
\description{
\code{fetchData} Fetches data from the API, handling pagination.
}
\examples{
fetchData(list(url = "https://api.datastream.org/v1/odata/v4/Records", qs = list(`$top` = 10)))
}
