% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastreamr.R
\name{fetchDataRateLimited}
\alias{fetchDataRateLimited}
\title{Fetch Data with Rate Limiting}
\usage{
fetchDataRateLimited(fetchOptions)
}
\arguments{
\item{fetchOptions}{A list of fetch options including URL and query string}
}
\value{
The API response
}
\description{
\code{fetchDataRateLimited} Fetches data from the API with rate limiting.
}
\examples{
fetchDataRateLimited(list(url = "https://api.datastream.org/v1/odata/v4/Records", qs = list(`$top` = 10)))
}
